fix papercuts
