+++ django_project/certification/templates/certificate_organisation/detail.html
@@ -81,7 +81,7 @@
         <div class="row" style="text-align: center; margin-top: 60px">
         <span class="glyphicon glyphicon-remove" style="color: red; font-size: 40pt"></span>
         <h3>Sorry, no certificate matching this id has been issued.</h3><br>
-        <a class="btn btn-default btn-mini" href="{% url 'validate-certificate' project_slug %}" style="font-weight: 600">Search another certificate</a>
+        <a class="btn btn-default btn-mini" href="{% url 'validate-certificate-organisation' project_slug %}" style="font-weight: 600">Search another certificate</a>
         </div>
     {% endif %}
