+++ django_project/certification/templates/certificate_organisation/validation.html
@@ -16,7 +16,7 @@
         $(function() {
             $("#submit-button").on("click", function() {
                 var query = $("#InputID").val();
-                location.href = '/en/' + '{{ project_slug }}' + '/organisationcertificate/' + query + '/detail/';
+                location.href = '/en/' + '{{ project_slug }}' + '/organisationcertificate/' + query + '/';
             });
         });
     </script>
