+++ django_project/certification/views/certificate_organisation.py
@@ -1,15 +1,139 @@
 # coding=utf-8
 import os
+from datetime import datetime
 from braces.views import LoginRequiredMixin
+from django.conf import settings
 from django.core.urlresolvers import reverse
 from django.http import HttpResponse, Http404
 from django.views.generic import CreateView, DetailView
+from reportlab.lib.utils import ImageReader
+from reportlab.lib.pagesizes import A4, landscape
+from reportlab.pdfgen import canvas
+from reportlab.pdfbase import pdfmetrics
+from reportlab.pdfbase.ttfonts import TTFont, TTFError
 from base.models.project import Project
-from base.views.certificate_preview import generate_certificate_pdf
 from certification.models.certifying_organisation import CertifyingOrganisation
 from certification.models.organisation_certificate import \
     CertifyingOrganisationCertificate
 from certification.forms import OrganisationCertificateForm
+
+
+def generate_certificate_pdf(
+        pathname, project, certifying_organisation, certificate, current_site):
+    """Create the PDF object, using the response object as its file."""
+
+    # Register new font
+    try:
+        font_folder = os.path.join(
+            settings.STATIC_ROOT, 'fonts/NotoSans-hinted')
+        ttf_file = os.path.join(font_folder, 'NotoSans-Bold.ttf')
+        pdfmetrics.registerFont(TTFont('Noto-bold', ttf_file))
+    except TTFError:
+        pass
+
+    page = canvas.Canvas(pathname, pagesize=landscape(A4))
+    width, height = A4
+    center = height * 0.5
+
+    if project.image_file:
+        project_logo = ImageReader(project.image_file)
+    else:
+        project_logo = None
+
+    if project.project_representative_signature:
+        project_representative_signature = \
+            ImageReader(project.project_representative_signature)
+    else:
+        project_representative_signature = None
+
+    if project.template_certifying_organisation_certificate:
+        background = \
+            ImageReader(project.template_certifying_organisation_certificate)
+    else:
+        background = None
+
+    # Certificate margin.
+    margin_right = height - 50
+    margin_left = 50
+    margin_bottom = 50
+    max_left = margin_right - 100
+
+    # Draw things on the PDF. Here's where the PDF generation happens.
+    # See the ReportLab documentation for the full list of functionality.
+    if background is not None:
+        page.drawImage(
+            background, 0, 0, height=width, width=height,
+            preserveAspectRatio=True, mask='auto')
+    page.setFillColorRGB(0.1, 0.1, 0.1)
+    page.setFont('Times-Roman', 18)
+
+    if project_logo is not None:
+        page.drawImage(
+            project_logo, 50, 450, width=100, height=100,
+            preserveAspectRatio=True, mask='auto')
+
+    page.setFont('Times-Roman', 12)
+    date_now = datetime.now()
+    str_date = date_now.strftime("%m/%d/%Y")
+    page.drawRightString(
+        margin_right, width - 50, 'Date issued: {}'.format(str_date))
+
+    try:
+        page.setFont('Noto-bold', 26)
+    except KeyError:
+        page.setFont('Times-Bold', 26)
+
+    page.drawCentredString(
+        center, 350, '{}'.format(certifying_organisation.name))
+    page.setFont('Times-Roman', 16)
+    page.drawCentredString(
+        center, 320,
+        'Is authorized to provide {} training and certification.'.format(
+            project.name))
+
+    page.setFillColorRGB(0.1, 0.1, 0.1)
+    if project_representative_signature is not None:
+        page.drawImage(
+            project_representative_signature,
+            (margin_right - 200), (margin_bottom + 70),
+            width=100,
+            height=70,
+            preserveAspectRatio=True,
+            anchor='s',
+            mask='auto')
+
+    page.setFont('Times-Italic', 12)
+    if project.project_representative:
+        page.drawCentredString(
+            (margin_right - 150), (margin_bottom + 60),
+            '{} {}'.format(
+                project.project_representative.first_name.encode('utf-8'),
+                project.project_representative.last_name.encode('utf-8')))
+    page.line(
+        (margin_right - 70), (margin_bottom + 55),
+        (margin_right - 230), (margin_bottom + 55))
+    page.setFont('Times-Roman', 13)
+    page.drawCentredString(
+        (margin_right - 150),
+        (margin_bottom + 40),
+        'Project Representative')
+
+    # Footnotes.
+    page.setFont('Times-Roman', 14)
+    page.drawString(
+        margin_left,
+        margin_bottom - 10,
+        'ID: {}'.format(certificate.certificateID))
+    page.setFont('Times-Roman', 8)
+    page.drawString(
+        margin_left, (margin_bottom - 20),
+        'You can verify this certificate by visiting '
+        'http://{}/en/{}/organisationcertificate/{}/.'
+        .format(current_site, project.slug, certificate.certificateID))
+
+    # Close the PDF object cleanly.
+    page.showPage()
+    page.save()
 
 
 class OrganisationCertificateCreateView(LoginRequiredMixin, CreateView):
@@ -140,8 +264,11 @@ class OrganisationCertificateDetailView(DetailView):
         context = super(
             OrganisationCertificateDetailView, self).get_context_data(**kwargs)
         context['project_slug'] = self.project_slug
-        context['history'] = \
-            context['certificate'].history.all().order_by('history_date')
+        try:
+            context['history'] = \
+                context['certificate'].history.all().order_by('history_date')
+        except KeyError:
+            pass
 
         if self.project_slug:
             context['the_project'] = \
